refactor: move navbar components
