+++ client/src/pages/Categories.js
@@ -1,5 +1,5 @@
 import React from "react";
-import Navbar from "../components/Navbar";
+import Navbar from "../components/navbar/Navbar";
 import Footer from "../components/Footer";
 import CategoryBox from "../components/categories/CategoryBox";
 import "../styles/categories.css"
